Renamespace factory
