+++ src/MessageDispatcher/DeliverEventMessageDispatcher.php
@@ -22,7 +22,7 @@ use App\Event\TestPassedEvent;
 use App\Event\TestStartedEvent;
 use App\Message\DeliverEventMessage;
 use App\Repository\JobRepository;
-use App\Services\WorkerEventFactory\EventHandler\WorkerEventFactory;
+use App\Services\WorkerEventFactory;
 use App\Services\WorkerEventStateMutator;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\Messenger\Envelope;
